requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
antiorm==1.2.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
bitarray==3.5.2
cffi==1.17.1
crc32c==2.7.1
cryptography==45.0.5
DateTime==5.5
db==0.1.1
frozenlist==1.7.0
greenlet==3.1.1
idna==3.10
multidict==6.4.4
pip==25.1.1
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic_core==2.27.2
pydantic==2.10.6
pydantic-settings==2.8.1
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2025.2
setuptools==65.5.0
SQLAlchemy==2.0.41
tonsdk==1.0.15
tvm-valuetypes==0.0.12
typing_extensions==4.12.2
websockets==15.0.1
wss==1.1
yarl==1.20.1
zope.interface==7.2
